Move Xcode target input files calculation from plugin to task
